experiments/006_chatty_bot/requirements.txt
# OpenAI SDK with Realtime API support
openai>=1.54.0

# WebSocket for OpenAI Realtime API
websockets>=12.0

# Audio processing
pyaudio>=0.2.14
numpy>=1.24.0

# Environment management
python-dotenv>=1.0.0

# Logging and monitoring
colorlog>=6.8.0

